--- internal/gossiper/consts.go
@@ -1,8 +1,0 @@
-// Copyright (C) 2024, Ava Labs, Inc. All rights reserved.
-// See the file LICENSE for licensing terms.
-
-package gossiper
-
-// initialCapacity is the initial size of a txs array we allocate when
-// unmarshaling a batch of txs.
-const initialCapacity = 1000
